angular material- crud coomplete

+++ src/app/person/person-list/person-list.component.ts
@@ -3,6 +3,9 @@ import {Person} from "../../domain/person";
 import {Router} from "@angular/router";
 import {PersonService} from "../../person/person.service";
 import {MatSnackBar} from "@angular/material/snack-bar";
+import {DialogData} from "../../dialog/dialog-data";
+import {DeleteConfirmDialogComponent} from "../../dialog/delete-confirm-dialog/delete-confirm-dialog.component";
+import {MatDialog} from "@angular/material/dialog";
 
 @Component({
   selector: 'app-person-list',
@@ -16,7 +19,8 @@ export class PersonListComponent implements OnInit {
 
   constructor(public router: Router,
               private personService: PersonService,
-              private snackBar: MatSnackBar) {
+              private snackBar: MatSnackBar,
+              public dialog: MatDialog) {
   }
 
   ngOnInit(): void {
@@ -32,20 +36,32 @@ export class PersonListComponent implements OnInit {
 
   goToDetail(p: Person | null) {
     if (p == null)
-      this.router.navigate(['detail']);
+      this.router.navigate(['persons', 'detail']);
     else
-      this.router.navigate(['detail', {id: p.id}]);
+      this.router.navigate(['persons', 'detail', {id: p.id}]);
   }
 
   delete(id: number) {
-    this.loading = true;
-    this.personService.delete(id).subscribe(p => {
-        this.findAll()
-        this.snackBar.open("La persona se elimino con exito", 'Éxito', {duration: 2000});
-      },
-      error => {
-        this.snackBar.open(error, "Error", {duration: 2000});
-        this.loading = false;
-      });
+    let dialogData = new DialogData(null,
+      "¿Está seguro que desea eliminar la persona?", "Confirmación para eliminar");
+    const dialogRef = this.dialog.open(DeleteConfirmDialogComponent, {
+      width: '300px',
+      height: 'auto',
+      minHeight: 200,
+      data: dialogData
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result.event == 'delete') {
+        this.loading = true;
+        this.personService.delete(id).subscribe(p => {
+            this.findAll()
+            this.snackBar.open("La persona se elimino con exito", 'Éxito', {duration: 2000});
+          },
+          error => {
+            this.snackBar.open(error, "Error", {duration: 2000});
+            this.loading = false;
+          });
+      }
+    });
   }
 }
